Refactor backport workflow structure
