+++ tools/backport/persistence.py
@@ -2,9 +2,10 @@
 
 # Standard library imports
 import json
-from dataclasses import asdict
+from dataclasses import asdict, fields
 from pathlib import Path
-from typing import Any
+from types import UnionType
+from typing import Any, cast, get_args, get_origin, get_type_hints
 
 # Bokeh imports
 from . import BackportError
@@ -76,130 +77,62 @@ def clear_plan(git: GitRepo) -> None:
 
 
 def _decode_plan(raw: dict[str, Any]) -> PlanState:
-    candidates = [_decode_candidate(item) for item in _list(raw, "candidates")]
-    return PlanState(
-        repository=_string(raw, "repository"),
-        version=_string(raw, "version"),
-        target_branch=_string(raw, "target_branch"),
-        branch=_string(raw, "branch"),
-        worktree=_string(raw, "worktree"),
-        base_sha=_sha(raw, "base_sha"),
-        candidates=candidates,
-        review_each=_boolean(raw, "review_each"),
-        remote=_string(raw, "remote"),
-        detached_worktree=_boolean(raw, "detached_worktree"),
-        pull_request_number=_optional_integer(raw, "pull_request_number"),
-        pull_request_url=_optional_string(raw, "pull_request_url"),
-        dedicated_commits=[_decode_dedicated_commit(item) for item in _list(raw, "dedicated_commits")],
-    )
+    state = _decode_dataclass(PlanState, raw)
+    for candidate in state.candidates:
+        if candidate.status not in VALID_CANDIDATE_STATES:
+            raise ValueError(f"unknown candidate state {candidate.status!r}")
+        _validate_sha("merge_sha", candidate.merge_sha)
+        _validate_sha("backport_sha", candidate.backport_sha)
+        _validate_sha("replay_sha", candidate.replay_sha)
+    for commit in state.dedicated_commits:
+        if commit.status not in {"pending", "applying", "conflict", "applied"}:
+            raise ValueError(f"unknown dedicated commit state {commit.status!r}")
+        _validate_sha("sha", commit.sha)
+        _validate_sha("backport_sha", commit.backport_sha)
+        _validate_sha("previous_head", commit.previous_head)
+    _validate_sha("base_sha", state.base_sha)
+    return state
 
 
-def _decode_candidate(raw: Any) -> Candidate:
+def _decode_dataclass[T](model: type[T], raw: Any) -> T:
     if not isinstance(raw, dict):
-        raise TypeError("candidate must be an object")
-    status = _string(raw, "status")
-    if status not in VALID_CANDIDATE_STATES:
-        raise ValueError(f"unknown candidate state {status!r}")
-    raw_issues = _list(raw, "issues")
-    if not all(isinstance(issue, dict) for issue in raw_issues):
-        raise TypeError("issues must contain only objects")
-    issues = [
-        IssueRef(
-            number=_integer(issue, "number"),
-            issue_type=_optional_string(issue, "issue_type"),
-        )
-        for issue in raw_issues
-    ]
-    return Candidate(
-        number=_integer(raw, "number"),
-        title=_string(raw, "title"),
-        url=_string(raw, "url"),
-        merged_at=_string(raw, "merged_at"),
-        merge_sha=_sha(raw, "merge_sha"),
-        base_branch=_string(raw, "base_branch"),
-        labels=_strings(raw, "labels"),
-        milestone=_optional_string(raw, "milestone"),
-        issues=issues,
-        status=status,
-        backport_sha=_optional_sha(raw, "backport_sha"),
-        replay_sha=_optional_sha(raw, "replay_sha"),
-        adapted=_boolean(raw, "adapted"),
-        conflict_files=_strings(raw, "conflict_files"),
-        reject_reason=_optional_string(raw, "reject_reason"),
-    )
+        raise TypeError(f"{model.__name__} must be an object")
+    model_fields = fields(cast(Any, model))
+    expected = {field.name for field in model_fields}
+    if set(raw) != expected:
+        missing = sorted(expected - set(raw))
+        extra = sorted(set(raw) - expected)
+        raise ValueError(f"{model.__name__} fields differ; missing={missing}, extra={extra}")
+    hints = get_type_hints(model)
+    values = {field.name: _decode_value(hints[field.name], raw[field.name], field.name) for field in model_fields}
+    return model(**values)
 
 
-def _decode_dedicated_commit(raw: Any) -> DedicatedCommit:
-    if not isinstance(raw, dict):
-        raise TypeError("dedicated commit must be an object")
-    status = _string(raw, "status")
-    if status not in {"pending", "applying", "conflict", "applied"}:
-        raise ValueError(f"unknown dedicated commit state {status!r}")
-    return DedicatedCommit(
-        sha=_sha(raw, "sha"),
-        subject=_string(raw, "subject"),
-        status=status,
-        backport_sha=_optional_sha(raw, "backport_sha"),
-        previous_head=_optional_sha(raw, "previous_head"),
-        conflict_files=_strings(raw, "conflict_files"),
-    )
-
-
-def _value(raw: dict[str, Any], key: str, expected: type) -> Any:
-    value = raw[key]
-    if not isinstance(value, expected):
-        raise TypeError(f"{key} must be {expected.__name__}")
+def _decode_value(annotation: Any, value: Any, name: str) -> Any:
+    origin = get_origin(annotation)
+    if origin is list:
+        if not isinstance(value, list):
+            raise TypeError(f"{name} must be list")
+        item_type = get_args(annotation)[0]
+        return [_decode_value(item_type, item, name) for item in value]
+    if origin is UnionType:
+        options = get_args(annotation)
+        if value is None and type(None) in options:
+            return None
+        concrete = [option for option in options if option is not type(None)]
+        if len(concrete) == 1:
+            return _decode_value(concrete[0], value, name)
+    if annotation in {PlanState, Candidate, DedicatedCommit, IssueRef}:
+        return _decode_dataclass(annotation, value)
+    if annotation is int:
+        if not isinstance(value, int) or isinstance(value, bool):
+            raise TypeError(f"{name} must be int")
+        return value
+    if annotation in {str, bool} and not isinstance(value, annotation):
+        raise TypeError(f"{name} must be {annotation.__name__}")
     return value
 
 
-def _string(raw: dict[str, Any], key: str) -> str:
-    return _value(raw, key, str)
-
-
-def _integer(raw: dict[str, Any], key: str) -> int:
-    value = _value(raw, key, int)
-    if isinstance(value, bool):
-        raise TypeError(f"{key} must be int")
-    return value
-
-
-def _optional_integer(raw: dict[str, Any], key: str) -> int | None:
-    value = raw[key]
-    if value is None:
-        return None
-    if not isinstance(value, int) or isinstance(value, bool):
-        raise TypeError(f"{key} must be int or null")
-    return value
-
-
-def _boolean(raw: dict[str, Any], key: str) -> bool:
-    return _value(raw, key, bool)
-
-
-def _list(raw: dict[str, Any], key: str) -> list[Any]:
-    return _value(raw, key, list)
-
-
-def _strings(raw: dict[str, Any], key: str) -> list[str]:
-    values = _list(raw, key)
-    if not all(isinstance(value, str) for value in values):
-        raise TypeError(f"{key} must contain only strings")
-    return values
-
-
-def _optional_string(raw: dict[str, Any], key: str) -> str | None:
-    value = raw[key]
-    if value is not None and not isinstance(value, str):
-        raise TypeError(f"{key} must be str or null")
-    return value
-
-
-def _sha(raw: dict[str, Any], key: str) -> str:
-    value = _string(raw, key)
-    if len(value) != 40 or any(character not in "0123456789abcdef" for character in value):
-        raise ValueError(f"{key} must be a full lowercase commit SHA")
-    return value
-
-
-def _optional_sha(raw: dict[str, Any], key: str) -> str | None:
-    return None if raw[key] is None else _sha(raw, key)
+def _validate_sha(name: str, value: str | None) -> None:
+    if value is not None and (len(value) != 40 or any(character not in "0123456789abcdef" for character in value)):
+        raise ValueError(f"{name} must be a full lowercase commit SHA")
