+++ tools/backport/publishing.py
@@ -7,11 +7,11 @@ from typing import Any
 from urllib.parse import quote
 
 # Bokeh imports
+from .aggregate import render_pr_body
 from .candidates import BACKPORT_LABEL, milestone_for_version
 from .git import GitRepo
 from .github import GitHubAPI, repository_path
 from .models import Candidate, PlanState
-from .planning import render_pr_body
 from .range_diffs import publish_range_diffs
